Revert "Fixed deprecated method in DatasetFileOperations"
This reverts commit a7006e7802e4f2b6fc0a3e6485c7141cd83e0219.

+++ WEB-INF/src/edu/unc/ceccr/workflows/datasets/DatasetFileOperations.java
@@ -949,8 +949,8 @@ public class DatasetFileOperations {
             for (int r = 0; r < rows; r++) {
                 row = sheet.getRow(r);
                 if (row != null) {
-                    cell1 = row.getCell(0);
-                    cell2 = row.getCell(1);
+                    cell1 = row.getCell((short) 0);
+                    cell2 = row.getCell((short) 1);
                     if (cell1.getCellType() == HSSFCell.CELL_TYPE_NUMERIC) {
                         sb.append((int) cell1.getNumericCellValue());
                     } else {
